+++ crates/client/db/src/class_db.rs
@@ -78,9 +78,10 @@ impl MadaraBackend {
         Ok(Some(info.class_info))
     }
 
-    pub fn contains_class(&self, id: &impl DbBlockIdResolvable, class_hash: &Felt) -> Result<bool, MadaraStorageError> {
-        // TODO(perf): make fast path, this only needs one db contains() call and no deserialization in most cases (block id pending/latest)
-        Ok(self.get_class_info(id, class_hash)?.is_some())
+    pub fn contains_class(&self, class_hash: &Felt) -> Result<bool, MadaraStorageError> {
+        let col = self.db.get_column(Column::ClassInfo);
+        let key_encoded = bincode::serialize(class_hash)?;
+        Ok(self.db.get_pinned_cf(&col, &key_encoded)?.is_some())
     }
 
     pub fn get_sierra_compiled(
@@ -123,12 +124,18 @@ impl MadaraBackend {
                 for converted_class in chunk {
                     let class_hash = converted_class.class_hash();
                     let key_bin = bincode::serialize(&class_hash)?;
-                    // TODO: find a way to avoid this allocation
-                    batch.put_cf(
-                        col,
-                        &key_bin,
-                        bincode::serialize(&ClassInfoWithBlockNumber { class_info: converted_class.info(), block_id })?,
-                    );
+                    // this is a patch because some legacy classes are declared multiple times
+                    if !self.contains_class(&class_hash)? {
+                        // TODO: find a way to avoid this allocation
+                        batch.put_cf(
+                            col,
+                            &key_bin,
+                            bincode::serialize(&ClassInfoWithBlockNumber {
+                                class_info: converted_class.info(),
+                                block_id,
+                            })?,
+                        );
+                    }
                 }
                 self.db.write_opt(batch, &writeopts)?;
                 Ok::<_, MadaraStorageError>(())