+++ crates/client/db/src/tests/test_open.rs
@@ -2,15 +2,12 @@ use super::common::*;
 use crate::DatabaseService;
 use mc_metrics::MetricsRegistry;
 use mp_chain_config::ChainConfig;
-use rstest::*;
 
-#[rstest]
 #[tokio::test]
 async fn test_open_db() {
     temp_db::temp_db().await;
 }
 
-#[rstest]
 #[tokio::test]
 async fn test_open_different_chain_id() {
     let temp_dir = tempfile::TempDir::new().unwrap();