+++ crates/client/sync/src/lib.rs
@@ -3,15 +3,16 @@ use anyhow::Context;
 use fetch::fetchers::FetchConfig;
 use mc_block_import::BlockImporter;
 use mc_db::MadaraBackend;
+use mc_gateway::client::builder::FeederClient;
 use mc_telemetry::TelemetryHandle;
-use mp_convert::ToFelt;
 use mp_exex::ExExManagerHandle;
-use starknet_providers::SequencerGatewayProvider;
+use reqwest::header::{HeaderName, HeaderValue};
 use std::{sync::Arc, time::Duration};
 
 pub mod fetch;
 pub mod l2;
 pub mod metrics;
+#[cfg(test)]
 pub mod tests;
 pub mod utils;
 
@@ -42,15 +43,13 @@ pub async fn sync(
 
     log::info!("⛓️  Starting L2 sync from block {}", starting_block);
 
-    let provider = SequencerGatewayProvider::new(
-        fetch_config.gateway.clone(),
-        fetch_config.feeder_gateway.clone(),
-        fetch_config.chain_id.to_felt(),
-    );
-    let provider = match &fetch_config.api_key {
-        Some(api_key) => provider.with_header("X-Throttling-Bypass".to_string(), api_key.clone()),
-        None => provider,
-    };
+    let mut provider = FeederClient::new(fetch_config.gateway, fetch_config.feeder_gateway);
+    if let Some(api_key) = fetch_config.api_key {
+        provider.add_header(
+            HeaderName::from_static("x-throttling-bypass"),
+            HeaderValue::from_str(&api_key).with_context(|| "Invalid API key format")?,
+        )
+    }
 
     l2::sync(
         backend,