+++ crates/tests/src/lib.rs
@@ -10,6 +10,7 @@ use std::ops::Range;
 use std::sync::Mutex;
 use std::{
     collections::HashMap,
+    env,
     future::Future,
     path::{Path, PathBuf},
     process::{Child, Command, Output, Stdio},
@@ -64,7 +65,7 @@ impl MadaraCmd {
                 res.error_for_status()?;
                 anyhow::Ok(())
             },
-            Duration::from_millis(1000),
+            Duration::from_millis(2000),
         )
         .await;
         self.ready = true;
@@ -85,7 +86,7 @@ impl MadaraCmd {
                     Err(err) => bail!(err),
                 }
             },
-            Duration::from_millis(5000),
+            Duration::from_millis(20000),
         )
         .await;
         self
@@ -133,7 +134,8 @@ pub fn get_port() -> MadaraPortNum {
     if let Some(el) = guard.to_reuse.pop() {
         return MadaraPortNum(el);
     }
-    MadaraPortNum(guard.next.next().expect("no more port to use"))
+    let port = guard.next.next().expect("no more port to use");
+    MadaraPortNum(port)
 }
 
 pub struct MadaraCmdBuilder {
@@ -170,22 +172,42 @@ impl MadaraCmdBuilder {
     }
 
     pub fn run(self) -> MadaraCmd {
+        let output = std::process::Command::new("cargo")
+            .arg("locate-project")
+            .arg("--workspace")
+            .arg("--message-format=plain")
+            .output()
+            .expect("Failed to execute command");
+
+        let cargo_toml_path = String::from_utf8(output.stdout).expect("Invalid UTF-8");
+        let project_root = PathBuf::from(cargo_toml_path.trim()).parent().unwrap().to_path_buf();
+
+        env::set_current_dir(&project_root).expect("Failed to set working directory");
         let target_bin = option_env!("COVERAGE_BIN").unwrap_or("./target/debug/madara");
         let target_bin = PathBuf::from_str(target_bin).expect("target bin is not a path");
         if !target_bin.exists() {
             panic!("No binary to run: {:?}", target_bin)
         }
 
+        // This is an optional argument to sync faster from the FGW if gateway_key is set
+        let gateway_key_arg =
+            std::env::var("GATEWAY_KEY").ok().map(|gateway_key| ["--gateway-key".into(), gateway_key]);
+
         let process = Command::new(target_bin)
             .envs(self.env)
-            .args(self.args.into_iter().chain([
-                "--telemetry-disabled".into(), // important: disable telemetry!!
-                "--no-prometheus".into(),
-                "--base-path".into(),
-                format!("{}", self.tempdir.as_ref().display()),
-                "--rpc-port".into(),
-                format!("{}", self.port.0),
-            ]))
+            .args(
+                self.args
+                    .into_iter()
+                    .chain([
+                        "--telemetry-disabled".into(), // important: disable telemetry!!
+                        "--no-prometheus".into(),
+                        "--base-path".into(),
+                        format!("{}", self.tempdir.as_ref().display()),
+                        "--rpc-port".into(),
+                        format!("{}", self.port.0),
+                    ])
+                    .chain(gateway_key_arg.into_iter().flatten()),
+            )
             .stdout(Stdio::piped())
             .spawn()
             .unwrap();
@@ -216,31 +238,27 @@ async fn madara_can_sync_a_few_blocks() {
     use starknet_core::types::{BlockHashAndNumber, Felt};
 
     let _ = env_logger::builder().is_test(true).try_init();
-    let mut cmd_builder = MadaraCmdBuilder::new().args([
+
+    let cmd_builder = MadaraCmdBuilder::new().args([
         "--full",
         "--network",
         "sepolia",
         "--no-sync-polling",
         "--n-blocks-to-sync",
-        "20",
+        "10",
         "--no-l1-sync",
     ]);
 
-    // This is an optional argument to sync faster from the FGW if gateway_key is set
-    if let Ok(gateway_key) = std::env::var("GATEWAY_KEY") {
-        cmd_builder = cmd_builder.args(["--gateway-key", &gateway_key]);
-    }
-
     let mut node = cmd_builder.run();
     node.wait_for_ready().await;
-    node.wait_for_sync_to(19).await;
+    node.wait_for_sync_to(9).await;
 
     assert_eq!(
         node.json_rpc().block_hash_and_number().await.unwrap(),
         BlockHashAndNumber {
-            // https://sepolia.voyager.online/block/19
-            block_hash: Felt::from_hex_unchecked("0x4177d1ba942a4ab94f86a476c06f0f9e02363ad410cdf177c54064788c9bcb5"),
-            block_number: 19
+            // https://sepolia.voyager.online/block/0x4174555d24718e8225a3d536ca96d2c4cc8a31bff6a6c758ab84a16a9e92d6c
+            block_hash: Felt::from_hex_unchecked("0x4174555d24718e8225a3d536ca96d2c4cc8a31bff6a6c758ab84a16a9e92d6c"),
+            block_number: 9
         }
     );
 }