+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - feat: `exex_pragma_dispatch` implementation
 - feat: Madara ExExs proof of concept
+- fix(sync): pending block retrying mechanism
+- fix:(tests): Add testing feature to mc-db dev dependency (#294)
+- feat: new crate gateway client & server
+- test: Starknet-js basic tests added
 - test: add block conversion task test
 - fix(docs): updated readme and fixed launcher
 - fix(ci): added gateway key to fix rate limit on tests